+++ BM_API/Repositories/RepositoryInterfaces/ICompanyEmployeeRepository.cs
@@ -7,7 +7,7 @@ namespace BM_API.Repositories.RepositoryInterfaces
         Task<CompanyEmployee> GetCompanyEmployeeByEmployeeId(Guid id);
         Task<ICollection<CompanyEmployee>> GetEmployeesByCompanyAsync(Guid companyId);
         Task<ICollection<Employee>> GetEmployeesWithEndDateAfterDateAsync(Guid companyId, DateTime date);
-        Task<decimal> GetSumOfSalariesAsync(Guid companyId, DateTime date);
+        Task<double> GetSumOfSalariesAsync(Guid companyId, DateTime date);
         Task<ICollection<Employee>> GetEmployeesBirthdaysForAMonthAsync(Guid companyId);
     }
 }
